volumes/requirements/base.txt
# Base frameworks
Django==3.1

# Database apps
psycopg
psycopg2-binary

# image & files apps
Pillow
django-imagekit

#template apps
django-compressor
cssmin
jsmin
django-crispy-forms==1.9.2

# Model tool apps
django-phone-field
django-ckeditor
django-phonenumber-field
Babel
phonenumbers

# Other Apps
python-decouple
gunicorn
django-debug-toolbar

